\name{dynsbm-package}
\alias{dynsbm-package}
\alias{dynsbm}
\docType{package}
\title{
Dynamic stochastic block model estimation
}
\description{
Estimation of a model that combines a stochastic block model (SBM) for its static part with independent Markov chains for the evolution of the nodes groups through time
}
\details{
dynsbm is a R implementation of a model that combines a stochastic block model (SBM) for its static part with independent Markov chains for the evolution of the nodes groups through time. It deals with binary or weighted dynamic/temporal/evolving networks (with discrete or continuous edges).
}
\author{
Authors: Catherine Matias, Vincent Miele
  
Maintainer: Vincent Miele <vincent.miele@univ-lyon1.fr>
}
\references{
Catherine Matias and Vincent Miele, \emph{Statistical clustering of temporal networks through a dynamic stochastic block model}, Journal of the Royal Statistical Society: Series B (2017)
http://dx.doi.org/10.1111/rssb.12200
http://arxiv.org/abs/1506.07464

Vincent Miele and Catherine Matias, \emph{Revealing the hidden structure
  of dynamic ecological networks},  Royal Society Open Science (2017)
http://dx.doi.org/10.1098/rsos.170251
https://arxiv.org/abs/1701.01355
}
\keyword{ package }
\keyword{ cluster }
%\seealso{
%Optional links to other man pages, e.g. \code{\link[<pkg>:<pkg>-package]{<pkg>}}
%}
%\examples{
%}
