\name{simdataT5Q4N40binary}
\docType{data}
\alias{simdataT5Q4N40binary}
\title{Dataset generated by a dynamic stochastic block model - binary case}
\description{
5 snapshots of a network with 40 nodes with binary weights, generated  by a dynamic stochastic block model with 4 groups.
}
\usage{data(simdataT5Q4N40binary)}
\format{An array of size 5x40x40.}