\name{simdataT5Q4N40continuous}
\docType{data}
\alias{simdataT5Q4N40continuous}
\title{Dataset generated by a dynamic stochastic block model - continuous case}
\description{
5 snapshots of a network with 40 nodes with continuous weights, generated  by a dynamic stochastic block model with 4 groups.
}
\usage{data(simdataT5Q4N40continuous)}
\format{An array of size 5x40x40.}