\name{simdataT5Q4N40discrete}
\docType{data}
\alias{simdataT5Q4N40discrete}
\title{Dataset generated by a dynamic stochastic block model - discrete case}
\description{
5 snapshots of a network with 40 nodes with discrete weights (K=3 categories), generated  by a dynamic stochastic block model with 4 groups.
}
\usage{data(simdataT5Q4N40discrete)}
\format{An array of size 5x40x40.}